+++ tests/bookmarks/test_xso.py
@@ -147,3 +147,53 @@ class TestURL(unittest.TestCase):
         self.assertNotEqual(url, url2)
         self.assertNotEqual(url, url3)
         self.assertNotEqual(url, conf)
+
+
+@bookmark_xso.as_bookmark_class
+class CustomBookmark(bookmark_xso.Bookmark):
+    TAG = ("urn:example:bookmark", "bookmark")
+
+    def __init__(self, name, contents):
+        self.name = name
+        self.contents = contents
+
+    name = aioxmpp.xso.Attr("name")
+    contents = aioxmpp.xso.Attr("contents")
+
+    @property
+    def primary(self):
+        return self.contents
+
+    @property
+    def secondary(self):
+        return (self.name,)
+
+
+class TestCustomBookmark(unittest.TestCase):
+
+    def test_registered(self):
+        self.assertIn(
+            CustomBookmark, bookmark_xso.Storage.bookmarks._classes
+        )
+
+    def test_non_Bookmarks_fail(self):
+        with self.assertRaises(TypeError):
+
+            @bookmark_xso.as_bookmark_class
+            class CustomBookmark(aioxmpp.xso.XSO):
+                TAG = ("urn:example:bookmark", "bookmark")
+
+                def __init__(self, name, contents):
+                    self.name = name
+                    self.contents = contents
+
+                name = aioxmpp.xso.Attr("name")
+                contents = aioxmpp.xso.Attr("contents")
+
+                @property
+                def primary(self):
+                    return self.contents
+
+                @property
+                def secondary(self):
+                    return (self.name,)